Fix passing a float for appropriate MXNet operations
When assigning each element of an n-dimensional array from a scalar, the
scalar should be of type float according to the MXNet documentation for
the operator `_set_value`. Also the operations with a scalar right-hand
side like `_plus_scalar`, `_minus_scalar`, `_mul_scalar`, `_div_scalar`,
etc. require the scalar to be of type float. This change replaces the
general type `T` with `float` for these cases.