Synchronize read access via `NDArray.data` and forbid writes
This change synchronizes for reads when calling `data` on n-dimensional
arrays. Note, this is only when you call `data`. When reusing the slice
for later access no synchronization is performed. User code must be
careful to perform manual synchronization via `NDArray.waitToRead` in
this case if necessary. Of course, you can always call `data` again
which performs the necessary synchronization.

This change also forbids writing to slice obtained via `data`. This is
achieved by returning a slice of type `Const!(T)[]`. Since writes cannot
be performed there is no need to synchronize them. The synchronized
`copyFrom` method can be used in this case to write all elements in a
synchronized manner.

In summary, this change makes `data` return a read-only slice of the
data elements of an n-dimensional array. Returning a read-only slice
makes this commit a breaking one. This avoids the need for
synchronization when writing elements via the slice obtained by `data`.
Read access is synchronized when executing the method.

Fixes sociomantic-tsunami/dmxnet#53.