Enable automatic minor update PRs for makd and beaver
This change enables providing automatic update PRs to minor versions for
makd and beaver. For all other submodules only patch versions are
allowed. This is enforced by explicitly specifying the default for
neptune libraries, which is patch.

These changes should result in an automatic PR for updating makd to
v2.2.0 and beaver to v0.6.0.