Update D bindings to MXNet version 1.2.0
This change updates the generated D bindings for the MXNet C interface.
The old bindings were generated from MXNet v1.1.0. This update provides
the changes for MXNet 1.2.0. The changes backward compatible except for
the changes for the profiling functions `MXSetProfilerConfig` and
`MXDumpProfile`. The dmxnet itself does not make use of these functions
but downstream may.